Multiple of 3,5 and 15
Write a program that loops through the numbers 1 through 100. Each
number should be printed to the console, using console.log(). However,
if the number is a multiple of 3, don't print the number, instead print
the word "fizz". If the number is a multiple of 5, print "buzz" instead
of the number. If it is a multiple of 3 and a multiple of 5, print
"fizzbuzz" instead of the number.